Melhora o tratamento de erros na exclusão de um usuário caso ele não exista na base de dados

+++ src/main/java/com/kleyton/gerenciador_de_tarefas_api/service/UsuarioService.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.stereotype.Service;
 
+import com.kleyton.gerenciador_de_tarefas_api.exceptions.AcessoNegadoException;
 import com.kleyton.gerenciador_de_tarefas_api.model.Usuario;
 import com.kleyton.gerenciador_de_tarefas_api.repositories.UsuarioRepository;
 
@@ -35,8 +36,10 @@ public class UsuarioService {
 	}
 
 	public void deletaUsuario(Long idUsuario) {
-		Usuario usuarioExistente = usuarioRepository.findById(idUsuario)
-				.orElseThrow(() -> new ResourceNotFoundException("Usuário não encontrado com ID: " + idUsuario));
+		if (usuarioRepository.existsById(idUsuario) == false) {
+			throw new AcessoNegadoException("Usuário não encontrado com ID: " + idUsuario);
+		}
+
 		usuarioRepository.deleteById(idUsuario);
 	}
 }
